17th push 03:03
17th push

+++ src/Services/crudPublication.java
@@ -7,6 +7,7 @@
 package Services;
 
 //import static controller.AffichageController.current_publication;
+import Controller.InscriController;
 import Entities.Publication;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -26,13 +27,15 @@ import Utiles.Basededonne;
 public class crudPublication {
      Connection con= Basededonne.getInstance().getConnection();
     private Statement ste;
-    
+    private ResultSet rs;
         public void ajouterPublication(Publication p) throws SQLException
         {
             //String req="insert into promotion (nom,date_debut,date_fin) values (?,?,?)";
-            PreparedStatement pre = (PreparedStatement) con.prepareStatement("INSERT INTO `publication`(`contenu`,`photo`,`datepublication`) VALUES (?,?,NOW())");
+            PreparedStatement pre = (PreparedStatement) con.prepareStatement("INSERT INTO `publication`(`contenu`,`photo`,`datepublication`,`id_produit`,`id_user`) VALUES (?,?,NOW(),?,?)");
             pre.setString(1, p.getContenu());
             pre.setString(2, p.getPhoto());
+            pre.setInt(3, p.getIdproduit());
+            pre.setInt(4, p.getIduser());
             //System.out.println(p.getContenu());
             pre.executeUpdate();
         }
@@ -67,9 +70,57 @@ public class crudPublication {
         
         public void supprimerPublication(int idsupp) throws SQLException
         {
-            PreparedStatement pre = (PreparedStatement) con.prepareStatement("delete from `publication` WHERE  id='"+idsupp+"'");
+            PreparedStatement pre = (PreparedStatement) con.prepareStatement("delete from `publication` WHERE  id='"+idsupp+"' and id_user='"+InscriController.current_user.getId()+"'");
             pre.executeUpdate();
         }
         
-
+        public List<String> afficher_produitParVille(String ville,int idCat) throws SQLException
+    {
+        String req="select produit.nom from produit inner join categorie on produit.id_categorie='"+idCat+"' and produit.ville='"+ville+"'";
+        ste=con.createStatement();
+        rs=ste.executeQuery(req);
+        List<String> listecat = new ArrayList<>();
+        while(rs.next())
+        {
+          listecat.add(rs.getString("produit.nom"));
+            System.out.println("aaaaaaaa");
+        }
+        return listecat ;
+        
+    }
+        
+        public List<Integer> ajouter_pubParId(String nom) throws SQLException
+    {
+        String req="select produit.id from produit inner join publication on produit.nom='"+nom+"'";
+        ste=con.createStatement();
+        rs=ste.executeQuery(req);
+        List<Integer> listecat = new ArrayList<>();
+        while(rs.next())
+        {
+          listecat.add(rs.getInt("produit.id"));
+            //System.out.println("aaaaaaaa");
+        }
+        return listecat ;
+        
+    }
+        
+        public List<Publication> afficherPublicationParId()
+        {
+            List<Publication> listePub = new ArrayList<>();
+         try {
+             //PreparedStatement pre = (PreparedStatement) con.prepareStatement("select * from `publication`");
+             
+             ste=con.createStatement();
+             ResultSet rs=ste.executeQuery("select * from publication where id_user='"+InscriController.current_user.getId()+"'");
+             while ( rs.next())
+             {
+                 Publication current= new Publication(rs.getInt("id"),rs.getString("contenu"),rs.getString("photo"),rs.getDate("datepublication"),rs.getInt("id_produit"));
+                 listePub.add(current);
+             }
+             
+         } catch (SQLException ex) {
+             Logger.getLogger(crudPublication.class.getName()).log(Level.SEVERE, null, ex);
+         }
+         return listePub;
+        }
 }
